pyproject.toml
[project]
name = "adk-skill-stream-ui"
version = "0.1.0"
description = "Agent platform with core runtime, workspace discovery, and React UI."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "google-adk>=1.18.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.9",
  "poethepoet>=0.32.0",
  "pytest>=8.3.5",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["api", "server"]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "workspace*"]
exclude = ["frontend*", "tests*"]
namespaces = true

[tool.poe.tasks]
backend = "uvicorn server:app --reload --reload-dir . --reload-include '*.py' --reload-exclude 'frontend/*' --reload-exclude '.venv/*' --reload-exclude '*/.venv/*' --reload-exclude 'venv/*' --reload-exclude '*/venv/*' --reload-exclude '*site-packages/*' --reload-exclude '__pycache__/*' --host 127.0.0.1 --port 8000"
frontend-install = "npm --prefix frontend install"
frontend = "npm --prefix frontend run dev -- --host 127.0.0.1 --port 3000"
test = "uv run python -m pytest"
stop = "zsh scripts/stop-dev.sh"
dev = { shell = "poe backend & BACKEND_PID=$!; poe frontend & FRONTEND_PID=$!; trap 'zsh scripts/stop-dev.sh >/dev/null 2>&1 || true; kill $BACKEND_PID $FRONTEND_PID 2>/dev/null || true' INT TERM EXIT; wait" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = ["--import-mode=importlib"]
